chore: encapsulate the snapshot retry domain logic
