+++ pkg/reconciler/backup/volumesnapshot/errors.go
@@ -17,15 +17,25 @@ limitations under the License.
 package volumesnapshot
 
 import (
+	"context"
+	"errors"
 	"regexp"
 	"strconv"
 	"strings"
+
+	apierrs "k8s.io/apimachinery/pkg/api/errors"
 )
 
 var (
 	retryableStatusCodes = []int{408, 429, 500, 502, 503, 504}
 	httpStatusCodeRegex  = regexp.MustCompile(`HTTPStatusCode:\s(\d{3})`)
 )
+
+// isErrorRetryable detects is an error is retryable or not
+func isErrorRetryable(err error) bool {
+	return apierrs.IsServerTimeout(err) || apierrs.IsConflict(err) || apierrs.IsInternalError(err) ||
+		errors.Is(err, context.DeadlineExceeded)
+}
 
 // isRetriableErrorMessage detects if a certain error message belongs
 // to a retriable error or not. This is obviously an heuristic but
